+++ REST API/Assignment2/IMDB/IMDB/Services/ReviewService.cs
@@ -1,4 +1,7 @@
-﻿using IMDB.Models;
+﻿using IMDB.CustomExceptions;
+using IMDB.Domain.Model;
+using IMDB.Domain.Request;
+using IMDB.Domain.Response;
 using IMDB.Repository;
 using IMDB.Repository.Interfaces;
 using IMDB.Services.Interfaces;
@@ -8,31 +11,78 @@ namespace IMDB.Services
     public class ReviewService:IReviewService
     {
         private readonly IReviewRepository _reviewRepository;
-        public ReviewService() { 
-            _reviewRepository=new ReviewRepository();
+        private readonly IMovieRepository _movieRepository;
+        private int _id = 0;
+        public ReviewService(IReviewRepository reviewRepository,IMovieRepository movieRepository) {
+            _reviewRepository = reviewRepository;
+            _movieRepository = movieRepository;
         }
 
-        public void Create(Review review)
+        public Review Create(ReviewRequest reviewRequest)
         {
-            _reviewRepository.Create(review);   
+            if (string.IsNullOrWhiteSpace(reviewRequest.Message)) throw new BadRequestException("Invalid Message");
+            if (reviewRequest.MovieId>_movieRepository.Get().Last().Id || reviewRequest.MovieId<=0) throw new BadRequestException("Invalid Id");
+            _id++;
+            return _reviewRepository.Create(new Review
+            {
+                Id = _id,
+                Message = reviewRequest.Message,
+                MovieId = reviewRequest.MovieId
+            });   
+            
         }
 
-        public List<Review> Get(int movieId) {
-            return _reviewRepository.Get(movieId);
-        }
-
-        public Review GetById(int movieId,int id)
+        public IList<ReviewResponse> Get()
         {
-            return _reviewRepository.GetById(movieId, id);
+            if (!_reviewRepository.Get().Any()) throw new BadRequestException("EMPTY");
+            return _reviewRepository.Get().Select(x=>new ReviewResponse
+            {
+                Id =x.Id,
+                Message = x.Message,
+                MovieId = x.MovieId
+            }).ToList();
         }
 
-        public void Update(Review review)
+        public IList<ReviewResponse> GetByMovieId(int movieId) {
+            if(!_reviewRepository.GetByMovieId(movieId).Any()) throw new BadRequestException("Invalid Message");
+            if (movieId>_movieRepository.Get().Last().Id ||  movieId<=0) throw new BadRequestException("Invalid Id");
+            var responseData=_reviewRepository.GetByMovieId(movieId);
+            return responseData.Select(x => new ReviewResponse
+            {
+                Id=x.Id,
+                MovieId=x.MovieId,
+                Message = x.Message,
+            }).ToList();
+        }
+
+        public ReviewResponse Get(int id)
         {
-            _reviewRepository.Update(review);
+            if (!_reviewRepository.Get().Any(x => x.Id == id)) throw new BadRequestException("Review not Found");
+            var responseData= _reviewRepository.Get(id);
+            return new ReviewResponse
+            {
+                Id=id,
+                Message=responseData.Message,
+                MovieId=responseData.MovieId
+            };
+        }
+
+        public void Update(int id,ReviewRequest reviewRequest)
+        {
+            if (string.IsNullOrWhiteSpace(reviewRequest.Message)) throw new BadRequestException("Invalid Message");
+            if (reviewRequest.MovieId>_movieRepository.Get().Last().Id || reviewRequest.MovieId<=0) throw new BadRequestException("Invalid Id");
+            _reviewRepository.Update(new Review
+            {
+                Id=id,
+                MovieId=reviewRequest.MovieId,
+                Message=reviewRequest.Message,
+            });
         }
 
         public void Delete(int id)
         {
+            if (id > _reviewRepository.Get().Last().Id || id <= 0) throw new BadRequestException("Invalid Id");
+            if (!_reviewRepository.Get().Any(x => x.Id == id)) throw new BadRequestException("Review not Found");
             _reviewRepository.Delete(id);
         }
     }
